Fixing issues with regex and type conversion steps
