+++ transform/steps.go
@@ -29,14 +29,6 @@ type FieldMapStep struct {
 	Sep    string `json:"sep"`
 }
 
-type RegexReplaceStep struct {
-	Column  string `json:"col"`
-	Regex   string `json:"regex"`
-	Replace string `json:"replace"`
-	Dest    string `json:"dst"`
-	reg     *regexp.Regexp
-}
-
 type AlleleIDStep struct {
 	Prefix         string `json:"prefix"`
 	Genome         string `json:"genome"`
@@ -111,20 +103,6 @@ func contains(s []string, q string) bool {
 		}
 	}
 	return false
-}
-
-func (re RegexReplaceStep) Run(i map[string]interface{}, task *pipeline.Task) map[string]interface{} {
-	col, _ := evaluate.ExpressionString(re.Column, task.Inputs, i)
-	replace, _ := evaluate.ExpressionString(re.Replace, task.Inputs, i)
-	dst, _ := evaluate.ExpressionString(re.Dest, task.Inputs, i)
-
-	o := re.reg.ReplaceAllString(col, replace)
-	z := map[string]interface{}{}
-	for x, y := range i {
-		z[x] = y
-	}
-	z[dst] = o
-	return z
 }
 
 func (al AlleleIDStep) Run(i map[string]interface{}, task *pipeline.Task) map[string]interface{} {
